+++ src/app/components/CardProposal.jsx
@@ -1,7 +1,7 @@
-import sponsor from '../../../public/aboutImages/conferenceRoom.jpg';
+import sponsor from '../../../public/conferenceRoom.jpg';
 import speaker from '../../../public/microphone.jpg';
 import Image from 'next/image';
-import style from '../../../styles/callToAction.module.css'
+import style from '../../../styles/proposal.module.css'
 import Btn from './Btn';
 
 const images = [
@@ -17,12 +17,12 @@ const images = [
     }
 ]
 
-export default function CardAction(props) {
+export default function ProposalSponsor(props) {
     return (
         <div>
             {
-                images.map((card) => {
-                    return <div className={style.action_Card}>
+                images.map((card, index) => {
+                    return <div  key={index} className={style.action_Card}>
                         <Image src={card.src} alt={card.description} width={150} />
                         <Btn action={card.text} />
                     </div>
