+++ src/app/components/TeamCard.jsx
@@ -5,12 +5,12 @@ export default function TeamCard(props) {
     return (
         <div >
             <div className={style.teamMember}>
-                <Image className={style.borderPic} src={props.urlPic} width={50} n height={50} />
+                <Image className={style.borderPic} src={props.urlPic} width={50} alt={props.description} height={50} />
                 <div className={style.memberData}>
                     <h5>{props.name}</h5>
                     <p>{props.role}</p>
                     <div>
-                        <Image src={props.networkUrl} width={15} height={15} />
+                        <Image src={props.networkUrl} alt={props.description} width={15} height={15} />
                     </div>
                 </div>
             </div>
