+++ src/app/components/Title.jsx
@@ -1,7 +1,7 @@
 import style from '../../../styles/main.module.css';
 
-export default function Title(props){
-    return(
+export default function Title(props) {
+    return (
         <h2 className={style.titles}>{props.title}</h2>
     )
 }
