+++ styles/proposal.module.css
